rag-backend/requirements.txt
fastapi
uvicorn
qdrant-client
google-generativeai
python-dotenv
fastapi-cors